sdk38: Rename pen.ts -> wallet.ts etc
